debug and some decs

+++ src/simulate/simulate_gc.cpp
@@ -185,8 +185,8 @@ namespace das
         }
         virtual void beforeTable ( Table * PT, TypeInfo * ti ) override {
             DataWalker::beforeTable(PT, ti);
-            auto tsize = (ti->firstType->size + ti->secondType->size + sizeof(uint32_t)) * PT->capacity;
-            DAS_ASSERT(tsize==(getTypeSize(ti->firstType)+getTypeSize(ti->secondType)+sizeof(uint32_t))*PT->capacity);
+            auto tsize = (ti->firstType->size + ti->secondType->size + sizeof(uint64_t)) * PT->capacity;
+            DAS_ASSERT(tsize==(getTypeSize(ti->firstType)+getTypeSize(ti->secondType)+sizeof(uint64_t))*PT->capacity);
             char * pa = PT->data;
             PtrRange rdata(pa, tsize);
             if ( reportHeap && tsize && markRange(rdata) ) {
@@ -675,7 +675,7 @@ namespace das
         }
         virtual void beforeTable ( Table * PT, TypeInfo * ti ) override {
             DataWalker::beforeTable(PT, ti);
-            PtrRange rdata(PT->data, (ti->firstType->size+ti->secondType->size+sizeof(uint32_t))*PT->capacity);
+            PtrRange rdata(PT->data, (ti->firstType->size+ti->secondType->size+sizeof(uint64_t))*PT->capacity);
             markAndPushRange(rdata);
         }
         virtual void afterTable ( Table * pa, TypeInfo * ti ) override {
